The conf file, and further setup improvements

+++ setup.sh
@@ -19,6 +19,17 @@ for entry in ${CONFARR[@]}; do
 		if [ "$target" == "$src" ]; then
 			echo "Link $dst up-to-date, skipping..."
 			continue
+		else
+			# File exists, is not a link to my dotfiles
+			echo "$dst exists and does not point to your dotfiles"
+			read -p "Overwrite? (y/n) " -n 1 -r
+			echo ""
+			if [[ ! "$REPLY" =~ [Yy]$ ]]; then
+				continue
+			else
+				mv "$dst" "$dst".bak
+				echo "Backing up to $dst.bak"
+			fi
 		fi
 	fi
 #	echo $dst
@@ -34,8 +45,9 @@ if hash openssl 2>/dev/null; then
 			decfile="./bash/.bash_aliases.d/${fname%.*}"
 			openssl des3 -d -salt -in "$encfile" -out "$decfile"
 			# If it failed due to wrong decryption key...
-			# the decrypted file is just crap
-			if [ "$?" -eq 1 ]; then
+			# the decrypted file is just crap. And file will
+			# not even exist if openssl got a sigint
+			if [ "$?" -eq 1 ] && [ -e "$decfile" ]; then
 				rm "$decfile"
 			fi
 		done
